File MetaData Completion and More Organized Program Files
The program Library object now has a function updateUnlabeledFiles() that will attempt to fill in all the missing information in the library using the internet and user input.
Also the program files are now seperated into individual objects for organization. The UI is based out of the main.py file. The user should run the main file to run the program.